+++ bindings/dart/flutter/colibri_flutter/example/linux/flutter/generated_plugin_registrant.cc
@@ -6,6 +6,10 @@
 
 #include "generated_plugin_registrant.h"
 
+#include <colibri_flutter/colibri_flutter_plugin.h>
 
 void fl_register_plugins(FlPluginRegistry* registry) {
+  g_autoptr(FlPluginRegistrar) colibri_flutter_registrar =
+      fl_plugin_registry_get_registrar_for_plugin(registry, "ColibriFlutterPlugin");
+  colibri_flutter_plugin_register_with_registrar(colibri_flutter_registrar);
 }