+++ bindings/dart/flutter/colibri_flutter/example/linux/flutter/generated_plugins.cmake
@@ -3,6 +3,7 @@
 #
 
 list(APPEND FLUTTER_PLUGIN_LIST
+  colibri_flutter
 )
 
 list(APPEND FLUTTER_FFI_PLUGIN_LIST