+++ bindings/dart/flutter/colibri_flutter/example/macos/Flutter/GeneratedPluginRegistrant.swift
@@ -5,6 +5,8 @@
 import FlutterMacOS
 import Foundation
 
+import colibri_flutter
 
 func RegisterGeneratedPlugins(registry: FlutterPluginRegistry) {
+  ColibriFlutterPlugin.register(with: registry.registrar(forPlugin: "ColibriFlutterPlugin"))
 }